Cleanup, move Event trait to own file
