+++ core/src/net/mod.rs
@@ -6,5 +6,7 @@ pub(crate) mod packet;
 pub use self::packet::Packet;
 
 pub(crate) mod server;
-pub use self::server::Events;
 pub use self::server::Server;
+
+pub(crate) mod event;
+pub use self::event::Events;
